MAINT: Reduce maximum code complexity to 30
Closes #3372.
